refactor: extract stripe helpers
